fix: improve resume file type validation and ensure deletion based on existence

+++ src/common/googleCloud.ts
@@ -81,7 +81,15 @@ export const deleteResumeByUserId = async (
     await Promise.all(
       // delete all possible accepted {id}.{ext} files uploaded by the user
       acceptedResumeExtensions.map((ext) =>
-        bucket.file(`${userId}.${ext}`).delete(),
+        bucket
+          .file(`${userId}.${ext}`)
+          .exists()
+          .then(([exists]) => {
+            if (exists) {
+              return bucket.file(`${userId}.${ext}`).delete();
+            }
+            return Promise.resolve(null);
+          }),
       ),
     );
